Unregister a dialog
When running on a "hubot-conversation" dialog and the user enters the last choice of it, that dialog doesn't get unregistered. So a possible robot.catchAll() listener in another script, it doesn't catch any messages even when the user is not running in a specific "hubot-conversation" dialog. Using the finish() function, solves that problem.